+++ sdk/src/main/java/ly/count/android/sdk/CountlyStarRating.java
@@ -43,7 +43,7 @@ public class CountlyStarRating {
 
     /**
      * Call to manually show star rating dialog
-     * @param context
+     * @param context android context
      * @param callback
      */
     public static void showStarRating(Context context, final CountlyStarRating.RatingCallback callback){
@@ -53,7 +53,7 @@ public class CountlyStarRating {
 
     /**
      * Method that created the star rating dialog
-     * @param context
+     * @param context android context
      * @param title
      * @param message
      * @param cancelText
@@ -230,7 +230,7 @@ public class CountlyStarRating {
 
     /**
      * Setting things that would be provided during initial config
-     * @param context
+     * @param context android context
      * @param limit limit for automatic rating
      * @param starRatingTextTitle provided title
      * @param starRatingTextMessage provided message
@@ -260,7 +260,7 @@ public class CountlyStarRating {
 
     /**
      * Returns a object with the loaded preferences
-     * @param context
+     * @param context android context
      * @return
      */
     private static StarRatingPreferences loadStarRatingPreferences(Context context) {
@@ -285,7 +285,7 @@ public class CountlyStarRating {
 
     /**
      * Save the star rating preferences object
-     * @param context
+     * @param context android context
      * @param srp
      */
     private static void saveStarRatingPreferences(Context context, StarRatingPreferences srp) {
@@ -295,7 +295,7 @@ public class CountlyStarRating {
 
     /**
      * Set if the star rating dialog should be shown automatically
-     * @param context
+     * @param context android context
      * @param shouldShow
      */
     public static void setShowDialogAutomatically(Context context, boolean shouldShow) {
@@ -308,7 +308,7 @@ public class CountlyStarRating {
      * Set if automatic star rating should be disabled for each new version.
      * By default automatic star rating will be shown for every new app version.
      * If this is set to true, star rating will be shown only once over apps lifetime
-     * @param context
+     * @param context android context
      * @param disableAsking if set true, will not show star rating for every new app version
      */
     public static void setStarRatingDisableAskingForEachAppVersion(Context context, boolean disableAsking) {
@@ -319,7 +319,7 @@ public class CountlyStarRating {
 
     /**
      * Register that a apps session has transpired. Will increase session counter and show automatic star rating if needed.
-     * @param context
+     * @param context android context
      * @param starRatingCallback
      */
     public static void registerAppSession(Context context, RatingCallback starRatingCallback) {
@@ -355,7 +355,7 @@ public class CountlyStarRating {
 
     /**
      * Returns how many sessions has star rating counted internally
-     * @param context
+     * @param context android context
      * @return
      */
     public static int getCurrentVersionsSessionCount(Context context){
@@ -365,7 +365,7 @@ public class CountlyStarRating {
 
     /**
      * Set the automatic star rating session count back to 0
-     * @param context
+     * @param context android context
      */
     public static void clearAutomaticStarRatingSessionCount(Context context){
         StarRatingPreferences srp = loadStarRatingPreferences(context);
@@ -375,7 +375,7 @@ public class CountlyStarRating {
 
     /**
      * Set if the star rating dialog is cancellable
-     * @param context
+     * @param context android context
      * @param isCancellable
      */
     public static void setIfRatingDialogIsCancellable(Context context, boolean isCancellable){
@@ -638,6 +638,7 @@ public class CountlyStarRating {
             DisplayMetrics metrics = new DisplayMetrics();
             activity.getWindowManager().getDefaultDisplay().getMetrics(metrics);
 
+            //noinspection RedundantIfStatement
             if (metrics.densityDpi == DisplayMetrics.DENSITY_DEFAULT
                     || metrics.densityDpi == DisplayMetrics.DENSITY_HIGH
                     || metrics.densityDpi == DisplayMetrics.DENSITY_MEDIUM
@@ -653,10 +654,12 @@ public class CountlyStarRating {
      * Used for detecting if device is a tv
      * @return
      */
+    @SuppressWarnings("RedundantIfStatement")
     protected static boolean isDeviceTv(Context context){
         final String TAG = "DeviceTypeRuntimeCheck";
 
         UiModeManager uiModeManager = (UiModeManager) context.getSystemService(UI_MODE_SERVICE);
+
         if (uiModeManager.getCurrentModeType() == Configuration.UI_MODE_TYPE_TELEVISION) {
             return true;
         } else {